Add unit tests for Server
This commit tests Server::create_stack_user() and
Server::rhsm_register().

For each function, we mock the method Server::run() and
then we run the tested function.

Thanks to the mock we can track the arguments on which
run() has been issued. Finally, we can assert that the
calls to run() match a pre-made list of arguments which
are stored in tests.commands.server.

Change-Id: Iaabee07d2bcd416ab4aaf8e69b10b74519af5846